feat: increase MAX_THREADS from 255 to 1024 via uint16_t n_threads (#349)
Change n_threads from uint8_t to uint16_t in ParseIndex to support
systems with more than 255 CPU cores. This is a breaking change for
the index file serialization format.

Changes:
- ParseIndex::n_threads changed from uint8_t to uint16_t
- Serialization format versioned (v2 uses uint16_t, v1 backward compat)
- MAX_THREADS in CLI increased from 255 to 1024
- All loop variables iterating over threads updated to uint16_t
- Tests added for large thread counts and legacy format reading

Fixes #143

+++ include/libvroom.h
@@ -1250,7 +1250,7 @@ public:
       if (!idx.n_indexes)
         return 0;
       size_t total = 0;
-      for (uint8_t t = 0; t < idx.n_threads; ++t) {
+      for (uint16_t t = 0; t < idx.n_threads; ++t) {
         total += idx.n_indexes[t];
       }
       return total;
